Fixing Issue #13 by decode result of strftime.
Issue #13: TODO extracting doesn't work

This bug appeared, if the user uses a locale with non-ASCII characters
in date / time texts. Since UTF-8 is (IMHO) the default encoding of
Sublime Text 2, explicitly decoding the result as UTF-8 should now
return a valid unicode string and not longer raise a
UnicodeDecodeError.